Move published message predicates to published events class CIRC-1025
